lib: move universe, observatory and  supercomputer.
